+++ DekorTrade/src/hu/dekortrade/server/jdo/Zarasfizetes.java
@@ -27,19 +27,19 @@ public class Zarasfizetes {
 	private String penztaros;
 
 	@Persistent
-	private Float fizet;
+	private Double fizet;
 
 	@Persistent
-	private Float fizeteur;
+	private Double fizeteur;
 
 	@Persistent
-	private Float fizetusd;
+	private Double fizetusd;
 
 	@Persistent
 	private Date datum;
 
 	public Zarasfizetes(String zaras, String cedula, String vevo, String tipus, String megjegyzes,
-			String penztaros, Float fizet, Float fizeteur, Float fizetusd,
+			String penztaros, Double fizet, Double fizeteur, Double fizetusd,
 			Date datum) {
 		this.zaras = zaras;
 		this.cedula = cedula;
@@ -93,27 +93,27 @@ public class Zarasfizetes {
 		this.penztaros = penztaros;
 	}
 
-	public Float getFizet() {
+	public Double getFizet() {
 		return fizet;
 	}
 
-	public void setFizet(Float fizet) {
+	public void setFizet(Double fizet) {
 		this.fizet = fizet;
 	}
 
-	public Float getFizeteur() {
+	public Double getFizeteur() {
 		return fizeteur;
 	}
 
-	public void setFizeteur(Float fizeteur) {
+	public void setFizeteur(Double fizeteur) {
 		this.fizeteur = fizeteur;
 	}
 
-	public Float getFizetusd() {
+	public Double getFizetusd() {
 		return fizetusd;
 	}
 
-	public void setFizetusd(Float fizetusd) {
+	public void setFizetusd(Double fizetusd) {
 		this.fizetusd = fizetusd;
 	}
