+++ DekorTrade/src/hu/dekortrade/shared/serialized/CedulacikkSer.java
@@ -20,17 +20,17 @@ public class CedulacikkSer implements IsSerializable {
 
 	private Integer darab;
 
-	private Float ar;
+	private Double ar;
 
-	private Float areur;
+	private Double areur;
 
-	private Float arusd;
+	private Double arusd;
 
-	private Float fizet;
+	private Double fizet;
 
-	private Float fizeteur;
+	private Double fizeteur;
 
-	private Float fizetusd;
+	private Double fizetusd;
 
 	public String getCedula() {
 		return cedula;
@@ -96,51 +96,51 @@ public class CedulacikkSer implements IsSerializable {
 		this.darab = darab;
 	}
 
-	public Float getAr() {
+	public Double getAr() {
 		return ar;
 	}
 
-	public void setAr(Float ar) {
+	public void setAr(Double ar) {
 		this.ar = ar;
 	}
 
-	public Float getAreur() {
+	public Double getAreur() {
 		return areur;
 	}
 
-	public void setAreur(Float areur) {
+	public void setAreur(Double areur) {
 		this.areur = areur;
 	}
 
-	public Float getArusd() {
+	public Double getArusd() {
 		return arusd;
 	}
 
-	public void setArusd(Float arusd) {
+	public void setArusd(Double arusd) {
 		this.arusd = arusd;
 	}
 
-	public Float getFizet() {
+	public Double getFizet() {
 		return fizet;
 	}
 
-	public void setFizet(Float fizet) {
+	public void setFizet(Double fizet) {
 		this.fizet = fizet;
 	}
 
-	public Float getFizeteur() {
+	public Double getFizeteur() {
 		return fizeteur;
 	}
 
-	public void setFizeteur(Float fizeteur) {
+	public void setFizeteur(Double fizeteur) {
 		this.fizeteur = fizeteur;
 	}
 
-	public Float getFizetusd() {
+	public Double getFizetusd() {
 		return fizetusd;
 	}
 
-	public void setFizetusd(Float fizetusd) {
+	public void setFizetusd(Double fizetusd) {
 		this.fizetusd = fizetusd;
 	}
