+++ DekorTrade/src/hu/dekortrade/shared/serialized/CikkSer.java
@@ -30,43 +30,43 @@ public class CikkSer implements IsSerializable {
 
 	private String vamtarifaszam;
 
-	private Float fob;
+	private Double fob;
 
-	private Float szallitas;
+	private Double szallitas;
 
-	private Float ddu;
+	private Double ddu;
 
-	private Float ersz;
+	private Double ersz;
 
-	private Float elorar;
+	private Double elorar;
 
-	private Float ujfob;
+	private Double ujfob;
 
-	private Float ujszallitas;
+	private Double ujszallitas;
 
-	private Float ujddu;
+	private Double ujddu;
 
-	private Float ujersz;
+	private Double ujersz;
 
-	private Float ujelorar;
+	private Double ujelorar;
 
-	private Float ar;
+	private Double ar;
 
-	private Float areur;
+	private Double areur;
 
-	private Float arszorzo;
+	private Double arszorzo;
 
 	private Integer kiskarton;
 
 	private Integer darab;
 
-	private Float terfogat;
+	private Double terfogat;
 
-	private Float terfogatlab;
+	private Double terfogatlab;
 
-	private Float bsuly;
+	private Double bsuly;
 
-	private Float nsuly;
+	private Double nsuly;
 
 	private String leiras;
 
@@ -174,107 +174,107 @@ public class CikkSer implements IsSerializable {
 		this.vamtarifaszam = vamtarifaszam;
 	}
 
-	public Float getFob() {
+	public Double getFob() {
 		return fob;
 	}
 
-	public void setFob(Float fob) {
+	public void setFob(Double fob) {
 		this.fob = fob;
 	}
 
-	public Float getSzallitas() {
+	public Double getSzallitas() {
 		return szallitas;
 	}
 
-	public void setSzallitas(Float szallitas) {
+	public void setSzallitas(Double szallitas) {
 		this.szallitas = szallitas;
 	}
 
-	public Float getDdu() {
+	public Double getDdu() {
 		return ddu;
 	}
 
-	public void setDdu(Float ddu) {
+	public void setDdu(Double ddu) {
 		this.ddu = ddu;
 	}
 
-	public Float getErsz() {
+	public Double getErsz() {
 		return ersz;
 	}
 
-	public void setErsz(Float ersz) {
+	public void setErsz(Double ersz) {
 		this.ersz = ersz;
 	}
 
-	public Float getElorar() {
+	public Double getElorar() {
 		return elorar;
 	}
 
-	public void setElorar(Float elorar) {
+	public void setElorar(Double elorar) {
 		this.elorar = elorar;
 	}
 
-	public Float getUjfob() {
+	public Double getUjfob() {
 		return ujfob;
 	}
 
-	public void setUjfob(Float ujfob) {
+	public void setUjfob(Double ujfob) {
 		this.ujfob = ujfob;
 	}
 
-	public Float getUjszallitas() {
+	public Double getUjszallitas() {
 		return ujszallitas;
 	}
 
-	public void setUjszallitas(Float ujszallitas) {
+	public void setUjszallitas(Double ujszallitas) {
 		this.ujszallitas = ujszallitas;
 	}
 
-	public Float getUjddu() {
+	public Double getUjddu() {
 		return ujddu;
 	}
 
-	public void setUjddu(Float ujddu) {
+	public void setUjddu(Double ujddu) {
 		this.ujddu = ujddu;
 	}
 
-	public Float getUjersz() {
+	public Double getUjersz() {
 		return ujersz;
 	}
 
-	public void setUjersz(Float ujersz) {
+	public void setUjersz(Double ujersz) {
 		this.ujersz = ujersz;
 	}
 
-	public Float getUjelorar() {
+	public Double getUjelorar() {
 		return ujelorar;
 	}
 
-	public void setUjelorar(Float ujelorar) {
+	public void setUjelorar(Double ujelorar) {
 		this.ujelorar = ujelorar;
 	}
 
-	public Float getAr() {
+	public Double getAr() {
 		return ar;
 	}
 
-	public void setAr(Float ar) {
+	public void setAr(Double ar) {
 		this.ar = ar;
 	}
 
-	public Float getAreur() {
+	public Double getAreur() {
 		return areur;
 	}
 
-	public void setAreur(Float areur) {
+	public void setAreur(Double areur) {
 		this.areur = areur;
 	}
 
-	public Float getArszorzo() {
+	public Double getArszorzo() {
 		return arszorzo;
 	}
 
-	public void setArszorzo(Float arszorzo) {
+	public void setArszorzo(Double arszorzo) {
 		this.arszorzo = arszorzo;
 	}
 
@@ -294,35 +294,35 @@ public class CikkSer implements IsSerializable {
 		this.darab = darab;
 	}
 
-	public Float getTerfogat() {
+	public Double getTerfogat() {
 		return terfogat;
 	}
 
-	public void setTerfogat(Float terfogat) {
+	public void setTerfogat(Double terfogat) {
 		this.terfogat = terfogat;
 	}
 
-	public Float getTerfogatlab() {
+	public Double getTerfogatlab() {
 		return terfogatlab;
 	}
 
-	public void setTerfogatlab(Float terfogatlab) {
+	public void setTerfogatlab(Double terfogatlab) {
 		this.terfogatlab = terfogatlab;
 	}
 
-	public Float getBsuly() {
+	public Double getBsuly() {
 		return bsuly;
 	}
 
-	public void setBsuly(Float bsuly) {
+	public void setBsuly(Double bsuly) {
 		this.bsuly = bsuly;
 	}
 
-	public Float getNsuly() {
+	public Double getNsuly() {
 		return nsuly;
 	}
 
-	public void setNsuly(Float nsuly) {
+	public void setNsuly(Double nsuly) {
 		this.nsuly = nsuly;
 	}
